Add build scripts and remove prebuilt frameworks from master branch
Create git tags and releases. Also make sure releases are generated in the right order
Also build static framework to add to release entry

+++ Example/Cronet/CNAppDelegate.m
@@ -18,7 +18,7 @@
 // modifying that.
 - (NSString*)currentNetLogFileName {
     return [NSString
-            stringWithFormat:@"cronet-consumer-net-log%lu.json", _counter];
+            stringWithFormat:@"cronet-consumer-net-log%lu.json", (unsigned long)_counter];
 }
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
